reorient_img/requirements.txt
backports.functools-lru-cache==1.5
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.16.0
Pillow==5.4.1
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
six==1.12.0
subprocess32==3.5.3
